+++ packages/account-tree-controller/src/backup-and-sync/user-storage/network-utils.test.ts
@@ -1,4 +1,4 @@
-import { executeWithRetry } from './network-utils';
+import { executeWithRetry } from './network-utils.js';
 
 describe('BackupAndSync - UserStorage - NetworkUtils', () => {
   beforeEach(() => {