+++ packages/accounts-controller/src/utils.ts
@@ -9,7 +9,7 @@ import { sha256 } from 'ethereum-cryptography/sha256';
 import type { V4Options } from 'uuid';
 import { v4 as uuid } from 'uuid';
 
-import type { AccountId } from './AccountsController';
+import type { AccountId } from './AccountsController.js';
 
 /**
  * Returns the name of the keyring type.