+++ packages/wallet/src/types.ts
@@ -9,7 +9,9 @@ import type { ApprovalControllerInstanceOptions } from './initialization/instanc
 import type { ConnectivityControllerInstanceOptions } from './initialization/instances/connectivity-controller/types.js';
 import type { KeyringControllerInstanceOptions } from './initialization/instances/keyring-controller/types.js';
 import type { NetworkControllerInstanceOptions } from './initialization/instances/network-controller/types.js';
+import type { PasskeyControllerInstanceOptions } from './initialization/instances/passkey-controller/types.js';
 import type { RemoteFeatureFlagControllerInstanceOptions } from './initialization/instances/remote-feature-flag-controller/types.js';
+import type { SeedlessOnboardingControllerInstanceOptions } from './initialization/instances/seedless-onboarding-controller/types.js';
 import type { StorageServiceInstanceOptions } from './initialization/instances/storage-service/types.js';
 import type { TransactionControllerInstanceOptions } from './initialization/instances/transaction-controller/types.js';
 import type { InitializationConfiguration } from './initialization/types.js';
@@ -32,4 +34,6 @@ export type InstanceSpecificOptions = {
   remoteFeatureFlagController: RemoteFeatureFlagControllerInstanceOptions;
   storageService: StorageServiceInstanceOptions;
   transactionController?: TransactionControllerInstanceOptions;
+  passkeyController?: PasskeyControllerInstanceOptions;
+  seedlessOnboardingController?: SeedlessOnboardingControllerInstanceOptions;
 };